alga4.32 - Implementando serviços REST de cidades.

+++ src/main/java/com/algaworks/algafood/infrastructure/repository/StateRepositoryImpl.java
@@ -1,12 +1,15 @@
 package com.algaworks.algafood.infrastructure.repository;
 
+import com.algaworks.algafood.domain.model.Kitchen;
 import com.algaworks.algafood.domain.model.State;
 import com.algaworks.algafood.domain.repository.StateRepository;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import java.util.List;
+import java.util.Objects;
 
 @Component
 public class StateRepositoryImpl implements StateRepository {
@@ -33,8 +36,11 @@ public class StateRepositoryImpl implements StateRepository {
 
     @Override
     @Transactional
-    public void delete(State state) {
-        state = search(state.getId());
+    public void delete(Long stateId) {
+        State state = search(stateId);
+        if(Objects.isNull(state)) {
+            throw new EmptyResultDataAccessException(1);
+        }
         entityManager.remove(state);
     }
 }
